move to use concurrent hash map
- instead of using a standard std::map for the output L1 indexing
queue, where we had to rely on condition variables and mutexes enforcing
correct read and writes of the L1 index map, I have tried two implementations
of concurrent (and lock-free) hash maps:

    1) Grampa implementation by Jeff Preshing
        (https://preshing.com/20160201/new-concurrent-hash-maps-for-cpp/)
    2) libcuckoo
        (https://github.com/efficient/libcuckoo)

Based on Preshing's benchmark tests, Grampa implementation is faster than
libcuckoo, but both are much faster even that tbb::concurrent_hash_map, which
is what FELIX and the sROD uses! Due to ease of use and a more thoreau documentation
level, I have at the moment opted for the libcuckoo implementation. One of the
main reasons why, is that libcuckoo has implemented methods to retrieve the iterators
(begin(), end(), etc) which, give that the hash map is key'd by the L1ID's
provide convenient methods to retrieve the next L1ID I want to pop off / and
or check. This method, perhaps naive/inefficient, suits the setup that we will
have since we will not have an indepedent TTC line, as with the sROD/FELIX
where a dedicated TTC e-link is providing the sROD with the L1As. In our setup
we do not have this, and the L1A information is direct from the data. Perhaps I
should ask Christos about implemeneting the TTC line.

Anyways, with the libcuckoo implementation in place I can see >=99% efficiency
for loopback data coming in from 3 independent UDP ports at a rate of 10 kHz. Nearing
the 100 kHz line things start to drop off, and we being to lose the sequence. With
a more robust accounting of the L1As, this could perhaps be avoided but in any
case 10kHz is a good ceiling to have given the dominant test cases for the sw.