fix race condition at builder startup
I was posting to the io_service that is shared between the listeners
and builders their callbacks. the listeners and builders own threads
start at different times, and as part of their startup they call the io_service
run() method in their threads and start doing things on the things.
so I removed the use of the io_service in the DataBuilder altogether
and simply call build() in its own thread. the io_service is only needed
for the DataListeners as they us the udp::sockets whose data reception
callbacks are handled by the io_service.